# Optimize analysis on multiple hosts