--- a/cli/analyze.go
+++ b/cli/analyze.go
@@ -24,6 +24,7 @@ import (
 	"io"
 	"os"
 	"path/filepath"
+	"sort"
 	"strings"
 	"time"
 
@@ -423,8 +424,10 @@ func writeSegs(ctx *cli.Context, wrSegs io.Writer, ops bench.Operations, allThre
 
 	// Write segments per endpoint
 	eps := ops.SortSplitByEndpoint()
+	epsSorted := stringKeysSorted(eps)
 	if details && len(eps) > 1 {
-		for _, ops := range eps {
+		for _, ep := range epsSorted {
+			ops := eps[ep]
 			segs := ops.Segment(bench.SegmentOptions{
 				From:           start,
 				PerSegDuration: aDur,
@@ -636,4 +639,14 @@ func checkAnalyze(ctx *cli.Context) {
 		err := errors.New("-analyze.dur cannot be 0")
 		fatal(probe.NewError(err), "Invalid -analyze.dur value")
 	}
+}
+
+// stringKeysSorted returns the keys as a sorted string slice.
+func stringKeysSorted[K string, V any](m map[K]V) []string {
+	keys := make([]string, 0, len(m))
+	for k := range m {
+		keys = append(keys, string(k))
+	}
+	sort.Strings(keys)
+	return keys
 }