Before: Duration: 189.16s, Total samples = 1489.30s (787.30%)
After: Duration: 97.52s, Total samples = 167.14s (171.39%)

50% wall time reduction, much bigger core reduction.

Results match.